+++ MyClaw/Objects/Officer.h
@@ -9,5 +9,5 @@ public:
 	Officer(const WwdObject& obj, Player* player);
 
 	Rectangle2D GetRect() override;
-	pair<Rectangle2D, int8_t> GetAttackRect() override;
+	pair<Rectangle2D, int> GetAttackRect() override;
 };
