+++ MyClaw/Objects/RobberThief.h
@@ -8,5 +8,5 @@ class RobberThief : public BaseEnemy
 public:
 	RobberThief(const WwdObject& obj, Player* player);
 
-	pair<Rectangle2D, int8_t> GetAttackRect() override;
+	pair<Rectangle2D, int> GetAttackRect() override;
 };
